Move common components to the `common` directory
