+++ src/components/common/links/external-link.tsx
@@ -1,17 +1,18 @@
 import React from 'react'
-import { Link } from 'react-router-dom'
-import { ForkAwesomeIcon, IconName } from '../../fork-awesome/fork-awesome-icon'
+import { ForkAwesomeIcon, IconName } from '../../../fork-awesome/fork-awesome-icon'
 import { ShowIf } from '../common/show-if'
 import { LinkWithTextProps } from './types'
 
-export const InternalLink: React.FC<LinkWithTextProps> = ({ href, text, icon, className = 'text-light' }) => {
+export const ExternalLink: React.FC<LinkWithTextProps> = ({ href, text, icon, className = 'text-light' }) => {
   return (
-    <Link to={href}
+    <a href={href}
+      target="_blank"
+      rel="noopener noreferrer"
       className={className}>
       <ShowIf condition={!!icon}>
         <ForkAwesomeIcon icon={icon as IconName} fixedWidth={true}/>&nbsp;
       </ShowIf>
       {text}
-    </Link>
+    </a>
   )
 }
