+++ src/components/common/links/internal-link.tsx
@@ -1,18 +1,17 @@
 import React from 'react'
-import { ForkAwesomeIcon, IconName } from '../../fork-awesome/fork-awesome-icon'
+import { Link } from 'react-router-dom'
+import { ForkAwesomeIcon, IconName } from '../../../fork-awesome/fork-awesome-icon'
 import { ShowIf } from '../common/show-if'
 import { LinkWithTextProps } from './types'
 
-export const ExternalLink: React.FC<LinkWithTextProps> = ({ href, text, icon, className = 'text-light' }) => {
+export const InternalLink: React.FC<LinkWithTextProps> = ({ href, text, icon, className = 'text-light' }) => {
   return (
-    <a href={href}
-      target="_blank"
-      rel="noopener noreferrer"
+    <Link to={href}
       className={className}>
       <ShowIf condition={!!icon}>
         <ForkAwesomeIcon icon={icon as IconName} fixedWidth={true}/>&nbsp;
       </ShowIf>
       {text}
-    </a>
+    </Link>
   )
 }
