+++ src/components/common/links/translated-external-link.tsx
