+++ src/components/editor/task-bar/help-button.tsx
@@ -2,7 +2,7 @@ import React, { Fragment, useState } from 'react'
 import { Button, Card, Col, Modal, Row, Table } from 'react-bootstrap'
 import { Trans, useTranslation } from 'react-i18next'
 import { ForkAwesomeIcon } from '../../../fork-awesome/fork-awesome-icon'
-import { TranslatedExternalLink } from '../../links/translated-external-link'
+import { TranslatedExternalLink } from '../../common/links/translated-external-link'
 
 export const HelpButton: React.FC = () => {
   const { t } = useTranslation()
