+++ src/components/landing/layout/version-info/version-info.tsx
@@ -6,7 +6,7 @@ import { Link } from 'react-router-dom'
 import { ApplicationState } from '../../../../redux'
 import frontendVersion from '../../../../version.json'
 import { ShowIf } from '../../../common/show-if'
-import { TranslatedExternalLink } from '../../../links/translated-external-link'
+import { TranslatedExternalLink } from '../../../common/links/translated-external-link'
 import { VersionInputField } from './version-input-field'
 
 export const VersionInfo: React.FC = () => {
