+++ src/components/landing/pages/history/history-content/history-content.tsx
@@ -1,8 +1,8 @@
 import React, { Fragment, useState } from 'react'
 import { Alert, Row } from 'react-bootstrap'
 import { Trans, useTranslation } from 'react-i18next'
-import { PagerPagination } from '../../../../pagination/pager-pagination'
 import { LocatedHistoryEntry } from '../history'
+import { PagerPagination } from '../../../../common/pagination/pager-pagination'
 import { HistoryCardList } from '../history-card/history-card-list'
 import { HistoryTable } from '../history-table/history-table'
 import { ViewStateEnum } from '../history-toolbar/history-toolbar'
