+++ src/utils/historyUtils.ts
@@ -1,7 +1,7 @@
 import moment from 'moment'
+import { SortModeEnum } from '../components/common/sort-button/sort-button'
 import { HistoryEntry, HistoryJson, LocatedHistoryEntry, Location } from '../components/landing/pages/history/history'
 import { HistoryToolbarState } from '../components/landing/pages/history/history-toolbar/history-toolbar'
-import { SortModeEnum } from '../components/sort-button/sort-button'
 
 export function collectEntries (localEntries: HistoryEntry[], remoteEntries: HistoryEntry[]): LocatedHistoryEntry[] {
   const locatedLocalEntries = locateEntries(localEntries, Location.LOCAL)
